Fix smallest and largest CBOR integer roundtrip (#58)

+++ src/lib.rs
@@ -130,7 +130,7 @@ fn decode_dag_cbor_to_pyobject<R: Read + Seek>(
     let major = decode::read_major(r)?;
     Ok(match major.kind() {
         MajorKind::UnsignedInt => decode::read_uint(r, major)?.into_pyobject(py)?.into(),
-        MajorKind::NegativeInt => (-1 - decode::read_uint(r, major)? as i64).into_pyobject(py)?.into(),
+        MajorKind::NegativeInt => (-1 - decode::read_uint(r, major)? as i128).into_pyobject(py)?.into(),
         MajorKind::ByteString => {
             let len = decode::read_uint(r, major)?;
             PyBytes::new(py, &decode::read_bytes(r, len)?).into_pyobject(py)?.into()
@@ -239,7 +239,7 @@ fn encode_dag_cbor_from_pyobject<'py, W: Write>(
 
         Ok(())
     } else if obj.is_instance_of::<PyInt>() {
-        let i: i64 = obj.extract()?;
+        let i: i128 = obj.extract()?;
 
         if i.is_negative() {
             encode::write_u64(w, MajorKind::NegativeInt, -(i + 1) as u64)?
